requirements.txt
diffusers==0.25.0
fastapi==0.109.0
Pillow==10.2.0
torch==2.1.2
uvicorn>=0.15.0,<0.16.0
-e ./utils

